Disable libffi 3.4's static trampolines feature
This new feature was introduced in libffi 3.4 and is known to break
some downstream packages, notably ghc and gobject-introspection.
Upstream also added a configure-time option in this release to
disable the feature until all downstream users can be fixed to be
compatible with it, at which point it can be safely re-enabled.

See also: https://github.com/libffi/libffi/pull/647
Signed-off-by: Dan Bryant <dan@deepwavedigital.com>